+++ packages/dnb-ui-lib/scripts/prepub/tasks/__tests__/convertSvgToJsx.test.js
@@ -19,7 +19,9 @@ beforeAll(async () => {
 
 describe('run convertSvgToJsx to convert ES6 to ES5', () => {
   it('has to convert correctly and match snapshot', async () => {
-    const { default: Svg } = await import('./test-files/dist/bell')
+    const { default: Svg } = await import(
+      path.resolve(__dirname, 'test-files/dist/bell.js')
+    )
     const ComponentWrap = mount(<Svg />)
     expect(toJson(ComponentWrap)).toMatchSnapshot()
   })